Refactor ROS2 composable node handling and introduce launchers
- Reorganized composable node actor and specifications into a new module structure for better maintainability.
- Introduced a new `launchers` package to handle different ROS2 launch mechanisms, including `composable_node`, `independent_node`, `node_container`, and `ros2_launcher`.
- Implemented `ComposableNodeActor` and `ComposableSpec` in `launchers/composable_node.py` to manage loading of composable nodes via the LoadNode service.
- Added `RosWorker` class in `ros_worker.py` to manage the rclpy node and LoadNode service clients.
- Created utility functions for building node specifications and handling parameters in the new launcher modules.
- Ensured backward compatibility by re-exporting necessary components from the previous structure.
